✨ Add Dash to laptop script
Looking up documentation for various programming tasks was arduous in
itself. We would have to open a browser, search for the correct article,
and find the information we needed. It would have been much better if
there was a single source for all documentation. We added Dash to the
laptop script to fulfil this purpose.