Add a `urlencode` alias
Before, if we wanted to URL encode a string, we would fire up a web
browser and find a site to do that for us. This was time-consuming when
we could run a little Ruby script on our local machines. We added a
`urlencode` alias to run a one-line Ruby script to URL encode a string.