Add a `mergepdf` alias
Before, if we wanted to merge any PDFs, we would have to launch Preview
and do a bunch of drag and dropping. This was laborious and involved us
using the mouse. Since Mac OS X Tiger, Apple included a script to do
this for us. We added a `mergepdf` alias to merge PDFs from the command
line.