Add `airport`
Before, we would use Apple's Wireless Diagnostics tool to help with
networking issues. The tool's UI could be more intuitive and are a
hassle. macOS has a hidden `airport` CLI that gives us access to the
same information. Unfortunately, Apple hid the executable somewhere in
the internals of macOS. We added an `airport` alias for easier access.