Add `puma-dev`
Before, it was challenging when we had several Rails apps running. They
would often all run on the same port. We would have to stop one app and
restart the other. This was an unnecessary overhead. We added `puma-dev`
to resolve the ports for us and give us a cool `.test` TLD.