Make `laptop` start a MySQL Docker container
Before, we had no way of having a consistent local MySQL setup. Setting
up a new machine, we were at the mercy of the latest version of MySQL,
and we would run into issues. To add consistency to our setup, we made
`laptop` always install MySQL 8 via a Docker container.